# entity-aggregation-service/src/main/java/sa/qiwa/cache/aggregate/ms/presentation/api/AggregateApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.0.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package sa.qiwa.cache.aggregate.ms.presentation.api;


import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Mono;

import jakarta.validation.Valid;
import jakarta.annotation.Generated;

import sa.qiwa.cache.aggregate.ms.application.delegator.AggregateApiDelegate;
import sa.qiwa.cache.aggregate.ms.domain.model.AggregateRequest;
import sa.qiwa.cache.aggregate.ms.domain.model.EntityName;
import sa.qiwa.cache.aggregate.ms.domain.model.AggregateResponse;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen")
@Validated

public interface AggregateApi {

    default AggregateApiDelegate getDelegate() {
        return new AggregateApiDelegate() {};
    }

    /**
     * POST /aggregate/{entityName} : Aggregation of searched results
     * This service will aggregate searched results
     *
     * @param entityName  (required)
     * @param aggregateRequest Will contain required aggregation operators and corresponding fields (required)
     * @return Success (status code 200)
     *         or Bad request (status code 400)
     *         or Entity not found (status code 404)
     *         or Validation error (status code 409)
     *         or Internal server error (status code 500)
     *         or The request method is not supported (Not Implemented) (status code 501)
     *         or Service unavailable due to invalid backend response(Bad Gateway) (status code 502)
     *         or Service unavailable (Under maintenance or overloaded) (status code 503)
     */
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/aggregate/{entityName}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default Mono<ResponseEntity<AggregateResponse>> aggregate(
         @PathVariable("entityName") EntityName entityName,
         @Valid @RequestBody AggregateRequest aggregateRequest,
         final ServerWebExchange exchange
    ) {
        System.out.println("dddddddddddddddddd");
        return getDelegate().aggregate(entityName, aggregateRequest, exchange);
    }

}